mods to color and font selector

+++ includes/support/tii-mapper.h
@@ -52,10 +52,10 @@ public:
 	                         const QString &, const QString &);
 	void	get_coordinates	(position &, float &,
 	                         const QString &, const QString &);
-	int	distance_2	(float, float, float, float);
-	int	distance	(position, position);
-	int	distance	(float, float, float, float);
-	int	corner		(position, position);
+	float	distance_2	(float, float, float, float);
+	float	distance	(position, position);
+	float	distance	(float, float, float, float);
+	float	corner		(position, position);
 	bool	is_black	(uint16_t, uint8_t, uint8_t);
 	void	set_black	(uint16_t, uint8_t, uint8_t);
